+++ packages/react/src/themes/greyVest.js
@@ -7,7 +7,7 @@ import { withStateLens } from '../utils/mobx-react-utils'
 import {
   Flex,
   TextHighlight,
-  FilteredPicker,
+  NestedPicker,
   ModalFilterAdder,
   TagsInput,
   FilterList as BaseFilterList,
@@ -440,6 +440,13 @@ export let GVStyle = () => (
         border: solid 1px #f1f1f1;
         padding: 5px;
       }
+      
+      .panel-tree-picker > div {
+        border-right: solid 1px #eef0f1;
+      }
+      .panel-tree-picker > div:last-child {
+        border-right: none;
+      }
     `}
   </style>
 )
@@ -539,25 +546,47 @@ let iconMap = {
 }
 let Icon = ({ icon }) => <Dynamic component={iconMap[icon]} />
 
+let AddLabel = (
+  <Flex style={{ justifyContent: 'space-between', alignItems: 'center' }}>
+    Add Custom Filter
+    <i className="material-icons" style={{ opacity: 0.4 }}>
+      filter_list
+    </i>
+  </Flex>
+)
+
+let FilterListItem = observer(
+  ({ active, disabled, hasChildren, children, ...props }) => (
+    <div
+      style={{
+        padding: '10px 40px',
+        cursor: 'pointer',
+        fontSize: 18,
+        background: active ? '#ebebeb' : '#fff',
+        color: disabled ? '#9b9b9b' : '#000',
+      }}
+      {...props}
+    >
+      {hasChildren ? (
+        <Flex style={{ alignItems: 'center' }}>
+          {children}
+          <i className="material-icons" style={{ fontSize: 20 }}>
+            chevron_right
+          </i>
+        </Flex>
+      ) : (
+        children
+      )}
+    </div>
+  )
+)
+
 export let Adder = ModalFilterAdder({
   Button,
   Input,
   Highlight,
-  Item: ListGroupItem,
-  label: (
-    <span
-      style={{
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-      }}
-    >
-      Add Custom Filter
-      <i className="material-icons" style={{ opacity: 0.4 }}>
-        filter_list
-      </i>
-    </span>
-  ),
+  Item: FilterListItem,
+  label: AddLabel,
 })
 
 export let PagerItem = observer(({ active, disabled, ...x }) => (
@@ -586,8 +615,8 @@ export let ExampleTypes = ExampleTypeConstructor({
   Checkbox,
   RadioList: ButtonRadio,
   Table,
-  FieldPicker: defaultProps({ Input, Highlight, Item: ListGroupItem })(
-    FilteredPicker
+  FieldPicker: defaultProps({ Input, Highlight, Item: FilterListItem })(
+    NestedPicker
   ),
   ListGroupItem,
   TagsInput: defaultProps({ TagComponent })(TagsInput),